backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
motor==3.6.0
pymongo[aws]==4.9.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
bcrypt==4.1.2
reportlab==4.0.9
aiofiles==23.2.1
email-validator==2.2.0
azure-ai-formrecognizer>=3.3
boto3>=1.28
pdf2image>=1.16
pillow>=10.0
httpx>=0.25
google-generativeai>=0.8.0
PyPDF2>=3.0.0
json-repair>=0.30.0
# penguin-ai-sdk installed from local wheel in Dockerfile
